fixed -wsign-compare warning
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4756 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/PointCloud.hpp
@@ -119,7 +119,7 @@ namespace geometry {
 	    void resize(int w, int h){
 	        _width = w;
 	        _height=h;
-	        if(_width*_height>_data.size())
+	        if(_width*_height> (int) _data.size())
 	            _data.resize(_width*_height);
 	    };
